change Makefiles, extended Button class
