+++ src/SequenceSimulation/Makefile
@@ -1,16 +1,18 @@
+include ../common.mk
+
 all: simulation_main
 
 simulation_main: simulation_main.o sequence.o generator.o
-	g++ -std=c++11 -Wall simulation_main.o sequence.o generator.o -o simulation_main
+	g++ $(CXXFLAGS) simulation_main.o sequence.o generator.o -o simulation_main
 
 simulation_main.o: simulation_main.cc sequence.h generator.h
-	g++ -std=c++11 -Wall -c simulation_main.cc
+	g++ $(CXXFLAGS) -c simulation_main.cc
 
 sequence.o: sequence.cc sequence.h
-	g++ -std=c++11 -Wall -c sequence.cc
+	g++ $(CXXFLAGS) -c sequence.cc
 
 generator.o: generator.cc generator.h
-	g++ -std=c++11 -Wall -c generator.cc
+	g++ $(CXXFLAGS) -c generator.cc
 
 clean:
 	rm -f simulation_main *.o
